requirements.txt
garminconnect>=0.2.38
click>=8.0
python-dotenv>=1.0
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
